phase-2/backend/requirements.txt
fastapi==0.115.0
sqlmodel==0.0.22
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
alembic==1.13.3
asyncpg==0.30.0
psycopg2==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
better-exceptions==0.3.3
pytest==8.3.3
pytest-asyncio==0.23.7
httpx==0.27.2
python-multipart==0.0.12
python-dotenv==1.0.1